backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic-settings>=2.7
python-multipart>=0.0.18
PyMuPDF>=1.25
Unidecode>=1.3
python-dotenv>=1.0
bcrypt>=4.2
PyJWT>=2.9
